re-split Cucumber test start methods.
