+++ build.xml
@@ -1642,6 +1642,9 @@
             <test name="jmri.RunCucumberTest">
                 <formatter usefile="no" type="brief"/>
             </test>
+            <test name="apps.RunCucumberTest">
+                <formatter usefile="no" type="brief"/>
+            </test>
         </junit>
         <fail>
             <condition>
